* in PacketTypeTest, renamed testChangeRateMessage() to testConsumerFlowTokenMessage() * fixed encoding in ConsumerFlowTokenMessageCodec

+++ src/main/org/jboss/messaging/core/remoting/codec/ConsumerFlowTokenMessageCodec.java
@@ -37,7 +37,7 @@ public class ConsumerFlowTokenMessageCodec extends AbstractPacketCodec<ConsumerF
    protected void encodeBody(ConsumerFlowTokenMessage message, RemotingBuffer out) throws Exception
    {
       out.putInt(INT_LENGTH);
-      out.putFloat(message.getTokens());
+      out.putInt(message.getTokens());
    }
 
    @Override
